kona: move read_tx_data into the batch module
read_tx_data is span-batch-specific — it size-bounds against
MAX_SPAN_BATCH_ELEMENTS and returns SpanBatchError — with a single caller in
decode_tx_data, so it doesn't belong in the general utils module. Relocate it
(with its EIP2718_MAX_TX_TYPE const and tests) into batch/transactions.rs, make
it private, and drop the crate re-export.

Add a doc comment describing its narrow responsibility — split an element into
its optional type prefix and RLP payload, validating only RLP structure — and
noting it does not yet discard invalid transaction types, which are rejected
downstream by the typed decoder in full_txs.

No behavior change.
